Created modules for each step of data processing. Created example file. Now it is possible to just import all needed modules, autimatically get data and recieve provisions index for the selected service.
